requirements.txt
numpy==2.1.2
pandas==2.2.3
Pillow==11.0.0
torch==2.4.1
torchvision==0.19.1
torchvision==0.20.0
ultralytics==8.3.22
